#THIS IS BAD
our whole for turret is based around positioning servo, but that won't work. gear ratio is 25:120, even a full rotation doesnt do much. we need to replace our whole code to be with a CR rotation, there's no other way i guarantee you unless we switch to using a motor, and we can't do that.
i made a shitload of changes